// frontend/src/app/containers/Ficha/store/types.ts
export enum Type {
    FETCH_FICHA_REQUEST = 'FETCH_FICHA_REQUEST',
    FETCH_FICHA_SUCCESS = 'FETCH_FICHA_SUCCESS',
    FETCH_FICHA_FAILURE = 'FETCH_FICHA_FAILURE',
    ADD_FICHA_REQUEST = 'ADD_FICHA_REQUEST',
    ADD_FICHA_SUCCESS = 'ADD_FICHA_SUCCESS',
    ADD_FICHA_FAILURE = 'ADD_FICHA_FAILURE',
    DELETE_FICHA_REQUEST = 'DELETE_FICHA_REQUEST',
    DELETE_FICHA_SUCCESS = 'DELETE_FICHA_SUCCESS',
    DELETE_FICHA_FAILURE = 'DELETE_FICHA_FAILURE',
    UPDATE_FICHA_REQUEST = 'UPDATE_FICHA_REQUEST',
    UPDATE_FICHA_SUCCESS = 'UPDATE_FICHA_SUCCESS',
    UPDATE_FICHA_FAILURE = 'UPDATE_FICHA_FAILURE',
    UPDATE_FICHA_STATUS_REQUEST = 'UPDATE_FICHA_STATUS_REQUEST',
    UPDATE_FICHA_STATUS_SUCCESS = 'UPDATE_FICHA_STATUS_SUCCESS',
    UPDATE_FICHA_STATUS_FAILURE = 'UPDATE_FICHA_STATUS_FAILURE',
  }